requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.9
django-allauth==0.39.1
django-cors-middleware==1.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.18
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
oauthlib==3.0.1
openapi-codec==1.3.2
packaging==19.0
pluggy==0.12.0
psycopg2-binary==2.8.6
py==1.8.0
pycodestyle==2.5.0
PyJWT==1.7.1
pyparsing==2.4.0
pytest==4.6.3
pytest-cov==2.7.1
pytest-django==3.5.0
python-decouple==3.4
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.4.1
uritemplate==3.0.0
urllib3==1.26.2
wcwidth==0.1.7
whitenoise==4.1.2
zipp==0.5.1
